Adding new Linen to company and its warhause done

+++ HotelLinenManagerV2.ApplicationServices/API/Domain/Requests/WarehauseDetails/CreateDetailsRequest.cs
@@ -7,6 +7,6 @@ namespace HotelLinenManagerV2.ApplicationServices.API.Domain.Requests.WarehauseD
     {
         public int? HotelLinenId { get; set; }
         public int? WarehauseId { get; set; }
-        public int Amount { get; set; }
+        public ushort Amount { get; set; }
     }
 }
